+++ CHANGES.md
@@ -68,6 +68,9 @@
   ([#165](https://github.com/flyingcircusio/batou/issues/165))
 
 
+- Do not write secrets files without recipient.
+  ([#184](https://github.com/flyingcircusio/batou/issues/184))
+
 2.3b1 (2021-05-21)
 ------------------